don't create symbol link when it exist.

+++ install.sh
@@ -6,7 +6,7 @@
 _dir="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
 
 # install patched fonts
-ln -s "$_dir"/fonts ~/.fonts
+[[ -L ~/.fonts ]] || ln -s "$_dir"/fonts ~/.fonts
 "$_dir"/fonts/install.sh
 
 # setup vim
